requirements.txt
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
Click==7.0
enum34==1.1.6
Flask==0.12.3
futures==3.2.0
google-api-core==1.10.0
google-auth==1.6.3
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.9
grpcio==1.20.1
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
joblib==0.13.2
MarkupSafe==1.1.0
numpy==1.16.1
pandas==0.20.2
protobuf==3.7.1
pyasn1==0.4.5
pyasn1-modules==0.2.5
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
rsa==4.0
scikit-learn==0.20.2
scipy==1.2.1
six==1.12.0
urllib3==1.24.2
Werkzeug==0.14.1
